Split utils, fixed dysplaying shortcuts
