+++ src/background/utils/clone.ts
@@ -1,0 +1,3 @@
+export function simpleClone(obj) {
+    return JSON.parse(JSON.stringify(obj));
+}
